+++ examples/multiple_storage_integrations/main.tf
@@ -1,10 +1,13 @@
-# Example: Multiple Snowflake Storage Integrations
-#
+# -----------------------------------------------------------------------------
+# Terraform Snowflake Storage Integration - Multiple Integrations Example
+# -----------------------------------------------------------------------------
 # This example demonstrates how to use the snowflake-storage-integration module
-# to create multiple Snowflake storage integrations using a map of configurations.
+# to create multiple Snowflake storage integrations using a map of
+# configurations.
+# -----------------------------------------------------------------------------
 
 module "storage_integrations" {
-  source = "../../modules/snowflake-storage-integration"
+  source = "../.."
 
   storage_integration_configs = var.storage_integration_configs
 }