+++ examples/single_storage_integration/main.tf
@@ -1,10 +1,12 @@
-# Example: Single Snowflake Storage Integration
-#
+# -----------------------------------------------------------------------------
+# Terraform Snowflake Storage Integration - Single Integration Example
+# -----------------------------------------------------------------------------
 # This example demonstrates how to use the snowflake-storage-integration module
 # to create a single Snowflake storage integration for AWS S3.
+# -----------------------------------------------------------------------------
 
 module "storage_integration" {
-  source = "../../modules/snowflake-storage-integration"
+  source = "../.."
 
   storage_integration_configs = var.storage_integration_configs
 }